rename test module
